Changed naming of classes and adjusted Some functions
